[norme] vm

+++ srcs/vm/instr/ld.c
@@ -6,7 +6,7 @@
 /*   By: briviere <briviere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/15 10:18:13 by briviere          #+#    #+#             */
-/*   Updated: 2018/04/16 14:57:46 by briviere         ###   ########.fr       */
+/*   Updated: 2018/04/16 16:48:19 by briviere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,7 @@ void	instr_ld(t_vm *vm, t_proc *proc)
 		proc->regs[reg] = instr.args[0].value.dir;
 	else if (instr.args[0].code == IND_CODE)
 	{
-		addr = proc->pc + ((int16_t)instr.args[0].value.ind % IDX_MOD) %
-			MEM_SIZE;
+		addr = proc->pc + (instr.args[0].value.ind % IDX_MOD) % MEM_SIZE;
 		proc->regs[reg] =
 			read_arena((t_arena_args){vm->arena, addr, REG_SIZE});
 	}
